pyproject.toml
[tool.poetry]
name = "repl_python3_SkribbleWordsBotDB"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.14.0"
discord = "^1.0.1"
flask = "^1.1.2"
replit = "^2.0.0"
pymongo = "^3.11.2"
dnspython = "^2.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
